videoに対応
