+++ src/features/ArticleDetailBody/hooks/renderBlock.tsx
@@ -19,6 +19,7 @@ import { QuotePresentation } from "../presentations/blocks/quote";
 import { TablePresentation } from "../presentations/blocks/table";
 import { ToDoPresentation } from "../presentations/blocks/toDo";
 import { TogglePresentation } from "../presentations/blocks/toggle";
+import { VideoPresentation } from "../presentations/blocks/video";
 
 export const renderBlock = async (block: Block, pageId: string) => {
   // notionのブロックの種類によって、表示を変える
@@ -101,30 +102,7 @@ export const renderBlock = async (block: Block, pageId: string) => {
         />
       );
     case "video":
-      if (block[type]?.type === "external") {
-        const youtubeUrl = block[type]?.external?.url;
-        if (!youtubeUrl) return null;
-        const youtubeEmbedUrl = youtubeUrl.replace("shorts", "embed").replace("watch?v=", "embed/");
-        return (
-          <iframe
-            key={id}
-            height={500}
-            src={youtubeEmbedUrl}
-            allowFullScreen
-            title={youtubeUrl}
-            allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
-          />
-        );
-      }
-
-      if (block[type]?.type === "file") {
-        return (
-          <video key={id} controls>
-            <source src={block[type]?.file?.url} type="video/mp4" />
-            Your browser does not support the video tag.
-          </video>
-        );
-      }
+      <VideoPresentation video={block[type]} id={id} />;
     default:
       return null;
   }
